Cleanup and fixes to networking: - Tidied logging and error messages - Handle disconnects properly - Tidied JsonMessage hierarchy - Improved HandshakeHandler ordering - Added slf4j-jdk dependency for flow loggers - Handle invalid opcodes, messages, handlers more cleanly

+++ src/main/java/net/glowstone/net/codec/handshake/HandshakeCodec.java
@@ -8,12 +8,11 @@ import net.glowstone.net.message.handshake.HandshakeMessage;
 import java.io.IOException;
 
 public final class HandshakeCodec implements Codec<HandshakeMessage> {
-
     @Override
     public HandshakeMessage decode(ByteBuf buffer) throws IOException {
         final int version = ByteBufUtils.readVarInt(buffer);
         final String address = ByteBufUtils.readUTF8(buffer);
-        final short port = (short) buffer.readUnsignedShort();
+        final int port = buffer.readUnsignedShort();
         final int state = ByteBufUtils.readVarInt(buffer);
 
         return new HandshakeMessage(version, address, port, state);
@@ -25,6 +24,5 @@ public final class HandshakeCodec implements Codec<HandshakeMessage> {
         ByteBufUtils.writeUTF8(buf, message.getAddress());
         buf.writeShort(message.getPort());
         buf.writeInt(message.getPort());
-
     }
 }
